refactoring, adds js docs
