+++ src/assistant-source.ts
@@ -31,11 +31,8 @@ export { injectionNames } from "./injection-names";
 
 // Export Mixin/Constructor interfaces
 export type Constructor<T> = new (...args: any[]) => T;
-export interface Mixin<T> {
-  new (...args: any[]): T;
-}
+export type Mixin<T> = new (...args: any[]) => T;
 
 // Export decorators
-export { stayInContext } from "./components/state-machine/decorators/stay-in-context-decorator";
-export { clearContext } from "./components/state-machine/decorators/clear-context-decorator";
+export { stayInContext, clearContext } from "./components/state-machine/decorators/context";
 export { filter } from "./components/state-machine/filter-decorator";
