--- src/components/state-machine/decorators/clear-context-decorator.ts
@@ -1,11 +1,0 @@
-import { ClearContextCallback } from "../public-interfaces";
-
-export const clearContextMetadataKey = Symbol("metadata-key: clearContext");
-
-export function clearContext(needsClear?: ClearContextCallback) {
-  const metadata = needsClear ? { clearContext: needsClear } : { clearContext: () => true };
-
-  return function(targetClass: any) {
-    Reflect.defineMetadata(clearContextMetadataKey, metadata, targetClass);
-  };
-}
